requirements.txt
sqlite3
nextcord
python-dotenv